Abstracts exclude mixin to it's own repo/pack
The reason for this is because of how Seed (currently) handles includePaths for sass. Things get tricky when there are multiple packs that rely on the same pack (e.g. seed-publish), but that shared pack has different dependency versions.

I experienced this when testing out the exclude mixin introduced in v0.0.3. It could not be found because packs built with seed-publish#0.0.2 didn't have exclude (and the inclusion of seed-publish#0.0.3 didn't make a different for Sass).